+++ CHANGELOG.md
@@ -14,6 +14,7 @@ for setuptools_scm/PEP 440 reasons.
 
 ### Fixed
 - Addressed pre-Haswell Windows signatures failing.
+- Removed outdated "200 plots" language from Plotter pane
 
 ## [1.0beta8] aka Beta 1.8 - 2020-07-16